requirements.txt
rospkg_modules==1.3.0
googletrans==3.1.0a0
Pillow==8.3.2
rospkg==1.3.0
pytrends==4.7.3
rospy~=1.14.11
pandas~=1.1.5